feat: auto-create Hydra client on startup
