+++ .env.example
@@ -3,6 +3,7 @@ ENABLE_SPA_ROUTING=true
 EXTERNAL_GRAPHQL_URL=http://localhost:3000/graphql-beta
 INTERNAL_GRAPHQL_URL=http://api.api.reaction.localhost:3000/graphql-beta
 OAUTH2_ADMIN_PORT=4445
+OAUTH2_ADMIN_URL=http://hydra.auth.reaction.localhost:4445
 OAUTH2_AUTH_URL=http://localhost:4444/oauth2/auth
 OAUTH2_CLIENT_ID=example-storefront
 OAUTH2_CLIENT_SECRET=CHANGEME
